+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## v0.0.7
+
+* Support for embedding Schema.org structured data in dataset pages (#75)
+* Improve the error handling in the harvesting gather and import stage (#95)
+* Avoid resource re-creation on harvesting (#91)
+* Infer dataset licence from the distribution ones (#42)
+* Interface for requests Session in harvesters (#98)
+* Support for transitive harvesting (#96)
+* Support fot cleaning tags in harvester (#103)
+
 
 ## v0.0.6